+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - conversion of microtonal `WesternPitch`
 - fermata attachment on multimeasure rest (disappeared before)
+- printing of duration lines for both natural harmonic items
 
 ### Added
 - `duration_line_engraver` and `prepare_for_duration_line_based_notation` arguments for `SequentialEventToAbjadVoice` for more flexible duration line based setups